Mon, Nov 20, 2023, 5:16 PM +05:00

+++ python/sentimental-readability/readability.py
@@ -29,11 +29,11 @@ def count_sentences(text):
 
 
 # Average # of letters per 100 words
-L = (letter_count/count_words(text)) * 100
+L = float((letter_count/(count_words(text)))) * 100
 print(f"L {L}")
 
 # Average sentences per 100 words
-S = (count_sentences(text)/count_words(text)) * 100
+S = float((count_sentences(text)/(count_words(text)))) * 100
 print(f"S {S}")
 
 # Coleman-Liau index formula
